requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
joblib==1.0.1
    # via scikit-learn
numpy==1.20.1
    # via
    #   scikit-learn
    #   scipy
pillow==8.1.2
    # via UNKNOWN (setup.py)
scikit-learn==0.24.1
    # via UNKNOWN (setup.py)
scipy==1.6.1
    # via scikit-learn
threadpoolctl==2.1.0
    # via scikit-learn
